refactor: split utils.py into core and gui modules
